requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.37.26
botocore==1.37.26
certifi==2025.1.31
click==8.1.8
distro==1.9.0
faiss-cpu==1.10.0
Flask==3.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.2.4
openai==1.70.0
packaging==24.2
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
